requirements.txt
PyQt5==5.15.7
python-chess==1.9.4
SQLAlchemy==1.4.41
requests==2.28.1